% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collatePAD}
\alias{collatePAD}
\title{Collating Well Locations in a Well PAD}
\usage{
collatePAD(X)
}
\arguments{
\item{X}{a data frame as generated by findPads function but with rows limited to a single PAD}
}
\value{
a data frame
}
\description{
function to collate the well locations of a single PAD
}
\author{
Adam Cagle
}
