% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collateWHT}
\alias{collateWHT}
\title{Collating Well Trajectories}
\usage{
collateWHT(WHT, dist = 250)
}
\arguments{
\item{WHT}{a data fram containing columns xs, ys, xb, and yb. All values should be in lat/long}

\item{dist}{a number specifying the threshold distance for surface locations to be grouped in
PADs. Units are in feet.}
}
\value{
a data frame
}
\description{
function to generate collated well trajectories from well surface and bottomhole locations
}
\details{
Latitude and longitude are assumed to be WGS84. WHT can contain extra columns, but columns named
xs (surface longitude), ys (surface latitude), xb (bottomhole longitude), and yb (bottomhole) latitude
are requried. Extra colums will be prserved and passed back to the user in the output.
}
\author{
Adam Cagle
}
