% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findPads}
\alias{findPads}
\title{Grouping Wells Into PADs}
\usage{
findPads(cWHTutm, dist = 100)
}
\arguments{
\item{cWHTutm}{a data frame as generated by by utmTransform function (UTM Coordinates)}
}
\value{
a data frame
}
\description{
This function uses clustering to find surface locations within a user specified threshold
distance, asigns these groups of wells to a PAD, and gives the pad an ID.
}
\author{
Adam Cagle
}
