% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAzimuth}
\alias{getAzimuth}
\title{Finding The Azimuth Along the collated Well Trajectory}
\usage{
getAzimuth(cWHT)
}
\arguments{
\item{X}{a data frame as generated by the getDistances function.}
}
\value{
a data frame which includes oLength and dLength in feet.
}
\description{
function to calculate oAximuth, the azimuth along the collated latteral.
}
\author{
Adam Cagle
}
