% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCOMs}
\alias{getCOMs}
\title{Finding Surface Location COMs For Each PAD/Branch}
\usage{
getCOMs(cWHTutm)
}
\arguments{
\item{X}{a data frame as generated by the collatePAD function (see details).}
}
\value{
a data frame
}
\description{
function to populate entire data frame with the final surface COM's grouped by pad and branch.
}
\details{
Before using getCOMs, the collatePAD function needs to be applied looping through the
data frame generated by findPads two times. The first itteration should be over "pad", and the
second itteration should be over pad and branch. Branches are generated when it's found that the
PAD is of the "dual pitchfork" variety.
}
\author{
Adam Cagle
}
