% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getDistances}
\alias{getDistances}
\title{Finding The Distances Along Well Trajectories}
\usage{
getDistances(cWHT)
}
\arguments{
\item{X}{a data frame as generated by the getCOMs function.}
}
\value{
a data frame which includes oLength and dLength in feet.
}
\description{
function to calculate oLength and dLength, the length along the collated latteral and the direct
patch from the surface to bottomhome locations respectively.
}
\author{
Adam Cagle
}
