% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getZoneUTM}
\alias{getZoneUTM}
\title{Finding the UTM Zone Number}
\usage{
getZoneUTM(longitude)
}
\arguments{
\item{longitude}{a vector of length n}
}
\value{
a vector of length n with UTM Zone numbers
}
\description{
function to calculate the UTM zone wich contains the longitude provided
}
\author{
Adam Cagle
}
