% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeTable}
\alias{makeTable}
\title{Creating a Collated WHT data frame}
\usage{
makeTable(WHT)
}
\arguments{
\item{WHT}{a data fram containing columns xs, ys, xb, and yb. All values should be in lat/long}
}
\value{
a data frame
}
\description{
This function creates a new data frame that will have containers for all of the outputs generated
by the collateWHT function.
}
\details{
Latitude and longitude are assumed to be WGS84. WHT can contain extra columns, but columns named
xs (surface longitude), ys (surface latitude), xb (bottomhole longitude), and yb (bottomhole) latitude
are requried. Extra colums will be prserved and passed back to the user in the output.
}
\author{
Adam Cagle
}
