% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapIt}
\alias{mapIt}
\title{Mapping Collated Well Trajectories
function to plot collated well trajectories on an interactive map.}
\usage{
mapIt(cWHT)
}
\arguments{
\item{cWHT}{a data frame as generated by the collateWHT function.}
}
\value{
a leflet map object
}
\description{
Mapping Collated Well Trajectories
function to plot collated well trajectories on an interactive map.
}
\author{
Adam Cagle
}
