% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{utmTransform}
\alias{utmTransform}
\title{Transforming Coordinates To and From lat-long and x-y}
\usage{
utmTransform(cWHT, from = "lat-long", to = "x-y")
}
\arguments{
\item{cWHT}{data frame as generated by the makeTable function}

\item{from}{string specifying CRS of coordinates in cWHT}

\item{to}{string specifying CRS of coordinates desired in output}
}
\value{
a data frame
}
\description{
function to transform to and from geographic lat-long to UTM x-y
}
\details{
from and two must be different from one another and can be either "lat-long" or "x-y". The
default is from lat-long to x-y.
}
\seealso{
\code{spTransform}
}
\author{
Adam Cagle
}
